Separate out the code to common to be used in other CLI tools
